Make a build which works with C89 on RISC OS.
This is a relatively simple change of the code to work with C89
compiler I use on RISC OS. This means that it has variable declarations
at the start of the blocks, and doesn't have the C99 types in stdbool.h
or stdint.h.

The RISC OS build will be added later.