Add the RISC OS build code.
The RISC OS build is pretty simple - it's just some Makefiles that
build the existing code and a small tool C file that can be built
to create sprite files or print the QR code as text.

There's a .robuild.yaml that can be used to perform the build of
the code on the build service.