Add border configuration; invert the colours.
The QR Code images that were being generated were not quite right.

The colours were the wrong way around - although it worked with some
decoders, it didn't work with them all, and was against the
specification apparently.

The images were also not being decoded correctly because there should
be a border around the outside of the images. This should be 4 modules
on each edge, although many decoders are more lax about decoding when
there is less border around them. The tool now defaults to 4 modules
around the outside which should make it decodable.